Model scenario ms values as int64, end to end
Copilot's tenth-round finding: maxScenarioMs (315,576,000,000) exceeds
MaxInt32 while the fixture structs and validator used platform-width
int, so GOARCH=386 fails to COMPILE at the comparison — and typing
only the constant would trade the compile error for a decode error,
since schema-valid values above MaxInt32 could not land in the int
structs. The ms values are now int64 end to end: the five config
durations, advance.Ms, the assertDelayMs envelope, checkScenarioMs,
and millis() itself, with the width rationale on the validator.

Red-proven with the 32-bit compile check: GOOS=linux GOARCH=386 go vet
against the previous code fails with "maxScenarioMs (untyped int
constant 315576000000) overflows int" at the cited line (darwin/386 is
not a valid pair, hence the linux GOOS); after the conversion — and
after typing the constant itself int64, which the first pass missed
because the untyped constant still defaulted to int inside
fmt.Errorf's variadic — the same vet exits 0.

Gates: GOOS=linux GOARCH=386 go vet clean; native suite green for both
eventfeed packages, plain and -race; gofmt clean.

+++ go/pkg/basecamp/eventfeed/scenario_harness_test.go
@@ -750,4 +750,4 @@ func (h *scenarioHarness) signalHandler(dispositions map[string]string) eventfee
 	}
 }
 
-func millis(ms int) time.Duration { return time.Duration(ms) * time.Millisecond }
+func millis(ms int64) time.Duration { return time.Duration(ms) * time.Millisecond }
